# swagger-debug.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PhysioGest API",
    "version": "1.0.0",
    "description": "Documentação da API PhysioGest"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/agendas": {
      "get": {
        "summary": "Lista todas as agendas",
        "tags": [
          "Agendas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de agendas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Agenda"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "post": {
        "summary": "Cria uma nova agenda",
        "tags": [
          "Agendas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "date",
                  "time",
                  "patientId",
                  "description",
                  "userId"
                ],
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "time": {
                    "type": "string"
                  },
                  "patientId": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agenda criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agenda"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/agendas/user/{userId}": {
      "get": {
        "summary": "Lista agendas por usuário",
        "tags": [
          "Agendas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de agendas do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Agenda"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/agendas/{id}": {
      "get": {
        "summary": "Busca uma agenda por ID",
        "tags": [
          "Agendas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da agenda"
          }
        ],
        "responses": {
          "200": {
            "description": "Agenda encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agenda"
                }
              }
            }
          },
          "404": {
            "description": "Agenda não encontrada"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "put": {
        "summary": "Atualiza uma agenda",
        "tags": [
          "Agendas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da agenda"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "time": {
                    "type": "string"
                  },
                  "patientId": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agenda atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agenda"
                }
              }
            }
          },
          "404": {
            "description": "Agenda não encontrada"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "delete": {
        "summary": "Deleta uma agenda",
        "tags": [
          "Agendas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da agenda"
          }
        ],
        "responses": {
          "200": {
            "description": "Agenda deletada"
          },
          "404": {
            "description": "Agenda não encontrada"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Agenda": {
        "type": "object",
        "required": [
          "date",
          "time",
          "patientId",
          "description",
          "userId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único da agenda"
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "Data da agenda"
          },
          "time": {
            "type": "string",
            "description": "Horário da agenda"
          },
          "patientId": {
            "type": "string",
            "description": "ID do paciente"
          },
          "description": {
            "type": "string",
            "description": "Descrição da agenda"
          },
          "userId": {
            "type": "string",
            "description": "ID do usuário que criou"
          }
        },
        "example": {
          "id": "60d5ecb74b24c72b8c8b4567",
          "date": "2023-12-25",
          "time": "14:30",
          "patientId": "60d5ecb74b24c72b8c8b4568",
          "description": "Consulta fisioterapêutica",
          "userId": "60d5ecb74b24c72b8c8b4569"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}